move hostproto to hostyarg
