+++ clean.sh
@@ -2,6 +2,6 @@
 
 rm -Rf bin
 rm -Rf build
-rm hostproto/hostproto
-rm hostproto/test.uf2
+rm hostyarg/hostyarg
+rm hostyarg/test.uf2
 rm -Rf cyarg/build
