+++ cyarg/files_littlefs.c
@@ -1,4 +1,4 @@
-#include "../hostproto/littlefs/lfs.h"
+#include "../hostyarg/littlefs/lfs.h"
 #include "pico_lfs_hal.h"
 #include "pico_flash_fs.h"
