--- hostproto/README.md
@@ -1,9 +1,0 @@
-# hostproto
-
-Host based tooling for creating and manipulating proto images for devices.
-
-Usage
-
-```
-./hostproto [bootdemo|addfile|format|ls]
-```
