--- hostproto/go.mod
@@ -1,3 +1,0 @@
-module github.com/jhmcaleely/proto-lang/hostproto
-
-go 1.23.4
