+++ hostyarg/go.mod
@@ -1,0 +1,3 @@
+module github.com/jhmcaleely/proto-lang/hostyarg
+
+go 1.23.4
